+++ cmd/repo/setup/command.go
@@ -1,0 +1,58 @@
+package setup
+
+import (
+	"io"
+	"os"
+
+	"github.com/giantswarm/microerror"
+	"github.com/sirupsen/logrus"
+	"github.com/spf13/cobra"
+)
+
+const (
+	name            = "setup"
+	description     = `Configure github repository`
+	longDescription = `Configure github repository with :
+  * settings
+  * permissions
+  * branch protection`
+)
+
+type Config struct {
+	Logger *logrus.Logger
+	Stderr io.Writer
+	Stdout io.Writer
+}
+
+func New(config Config) (*cobra.Command, error) {
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+	if config.Stderr == nil {
+		config.Stderr = os.Stderr
+	}
+	if config.Stdout == nil {
+		config.Stdout = os.Stdout
+	}
+
+	f := &flag{}
+
+	r := &runner{
+		flag:   f,
+		logger: config.Logger,
+		stderr: config.Stderr,
+		stdout: config.Stdout,
+	}
+
+	c := &cobra.Command{
+		Use:   name,
+		Short: description,
+		Long:  longDescription,
+		RunE:  r.Run,
+		Args:  cobra.ExactArgs(1),
+	}
+
+	f.Init(c)
+
+	return c, nil
+}